provide clarity to project structure, commit initial display drivers and code.
