+++ src/GUI_Paint.cpp
@@ -1,74 +1,5 @@
 /******************************************************************************
-* | File      	:   GUI_Paint.c
-* | Author      :   Waveshare electronics
-* | Function    :	Achieve drawing: draw points, lines, boxes, circles and
-*                   their size, solid dotted line, solid rectangle hollow
-*                   rectangle, solid circle hollow circle.
-* | Info        :
-*   Achieve display characters: Display a single character, string, number
-*   Achieve time display: adaptive size display time minutes and seconds
-*----------------
-* |	This version:   V3.1
-* | Date        :   2020-08-15
-* | Info        :
-* -----------------------------------------------------------------------------
-* V3.1(2020-08-15):
-* 1.Fix: 
-*       Paint_DrawNum
-*         Fixed a BUG where the Paint_DrawNum function failed to display 0
-* 2.Add： Paint_DrawFloatNum
-*     Can display FloatNum   
-*
-* -----------------------------------------------------------------------------
-* V3.0(2019-04-18):
-* 1.Change: 
-*    Paint_DrawPoint(..., DOT_STYLE DOT_STYLE)
-* => Paint_DrawPoint(..., DOT_STYLE Dot_Style)
-*    Paint_DrawLine(..., LINE_STYLE Line_Style, DOT_PIXEL Dot_Pixel)
-* => Paint_DrawLine(..., DOT_PIXEL Line_width, LINE_STYLE Line_Style)
-*    Paint_DrawRectangle(..., DRAW_FILL Filled, DOT_PIXEL Dot_Pixel)
-* => Paint_DrawRectangle(..., DOT_PIXEL Line_width, DRAW_FILL Draw_Fill)
-*    Paint_DrawCircle(..., DRAW_FILL Draw_Fill, DOT_PIXEL Dot_Pixel)
-* => Paint_DrawCircle(..., DOT_PIXEL Line_width, DRAW_FILL Draw_Filll)
-*
-* -----------------------------------------------------------------------------
-* V2.0(2018-11-15):
-* 1.add: Paint_NewImage()
-*    Create an image's properties
-* 2.add: Paint_SelectImage()
-*    Select the picture to be drawn
-* 3.add: Paint_SetRotate()
-*    Set the direction of the cache    
-* 4.add: Paint_RotateImage() 
-*    Can flip the picture, Support 0-360 degrees, 
-*    but only 90.180.270 rotation is better
-* 4.add: Paint_SetMirroring() 
-*    Can Mirroring the picture, horizontal, vertical, origin
-* 5.add: Paint_DrawString_CN() 
-*    Can display Chinese(GB1312)   
-*
-* ----------------------------------------------------------------------------- 
-* V1.0(2018-07-17):
-*   Create library
-*
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documnetation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to  whom the Software is
-* furished to do so, subject to the following conditions:
-*
-* The above copyright notice and this permission notice shall be included in
-* all copies or substantial portions of the Software.
-*
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS OR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-* THE SOFTWARE.
-*
+
 ******************************************************************************/
 #include "GUI_Paint.h"
 #include "DEV_Config.h"
@@ -76,6 +7,7 @@
 #include <stdlib.h>
 #include <string.h> //memset()
 #include <math.h>
+#include "debug.h"
 
 volatile PAINT Paint;
 
@@ -533,67 +465,6 @@ void Paint_DrawString_EN(UWORD Xstart, UWORD Ystart, const char * pString,
     Color_Background : Select the background color of the English character
     Color_Foreground : Select the foreground color of the English character
 ******************************************************************************/
-void Paint_DrawString_CN(UWORD Xstart, UWORD Ystart, const char * pString, cFONT* font, UWORD Color_Background, UWORD Color_Foreground)
-{
- const unsigned char* p_text = pString;
-
-  int refcolumn = Xstart;
-  int i, j, Num;
-  /* Send the string character by character on EPD */
-  while (*p_text != 0) {
-    if (*p_text < 0x7F) {                                  //ASCII
-      for (Num = 0; Num < font->size ; Num++) {
-        if (*p_text == pgm_read_byte(&font->table[Num].index[0])) {
-          const char* ptr = &font->table[Num].matrix[0];
-
-          for (j = 0; j < font->Height; j++) {
-            for (i = 0; i < font->Width; i++) {
-              if (pgm_read_byte(ptr) & (0x80 >> (i % 8))) {
-                Paint_SetPixel(refcolumn + i,Ystart + j, Color_Foreground);
-              }
-              if (i % 8 == 7) {
-                ptr++;
-              }
-            }
-            if (font->Width % 8 != 0) {
-              ptr++;
-            }
-          }
-          break;
-        }
-      }
-      /* Point on the next character */
-      p_text += 1;
-      /* Decrement the column position by 16 */
-      refcolumn += font->ASCII_Width;
-    } else {                                   //中文
-      for (Num = 0; Num < font->size ; Num++) {
-        if ((*p_text == pgm_read_byte(&font->table[Num].index[0])) && (*(p_text + 1) == pgm_read_byte(&font->table[Num].index[1])) && (*(p_text + 2) == pgm_read_byte(&font->table[Num].index[2]))) {
-          const char* ptr = &font->table[Num].matrix[0];
-
-          for (j = 0; j < font->Height; j++) {
-            for (i = 0; i < font->Width; i++) {
-              if (pgm_read_byte(ptr) & (0x80 >> (i % 8))) {
-                Paint_SetPixel(refcolumn + i,Ystart + j, Color_Foreground);
-              }
-              if (i % 8 == 7) {
-                ptr++;
-              }
-            }
-            if (font->Width % 8 != 0) {
-              ptr++;
-            }
-          }
-          break;
-        }
-      }
-      /* Point on the next character */
-      p_text += 3;
-      /* Decrement the column position by 16 */
-      refcolumn += font->Width;
-    }
-  }
-}
 
 
 /******************************************************************************
