+++ src/LCD_Driver.cpp
@@ -1,31 +1,14 @@
 /*****************************************************************************
-* | File      	:	LCD_Driver.c
-* | Author      :   Waveshare team
-* | Function    :   LCD driver
+* | File        :   LCD_Display.ino
+* | Author      :   Brett Nelson, @throneofshadow
+* | Function    :   Demo script to run basic functionality of LCD screen using an Arduino Uno/Nano
 * | Info        :
+*                Used to provide master control, drawing, and mode switching.
 *----------------
-* |	This version:   V1.0
-* | Date        :   2018-12-18
-* | Info        :   
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documnetation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to  whom the Software is
-# furished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS OR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-#
+* | This version:   V0.1
+* | Date        :   2024-7-March
+* | Info        : Written for YouSolar Inc research and engineering team for device-level displays used in Gen 3 PowerBlock.
+* 
 ******************************************************************************/
 #include "LCD_Driver.h"
 /*******************************************************************************
