+++ src/fonts.h
@@ -1,0 +1,42 @@
+/**
+  ******************************************************************************
+  ******************************************************************************
+  */
+
+/* Define to prevent recursive inclusion -------------------------------------*/
+#ifndef __FONTS_H
+#define __FONTS_H
+
+#define MAX_HEIGHT_FONT         48
+#define MAX_WIDTH_FONT          40
+#define OFFSET_BITMAP           
+
+#ifdef __cplusplus
+ extern "C" {
+#endif
+
+/* Includes ------------------------------------------------------------------*/
+#include <stdint.h>
+#include <avr/pgmspace.h>
+//ASCII
+typedef struct _tFont
+{    
+  const uint8_t *table;
+  uint16_t Width;
+  uint16_t Height;
+  
+} sFONT;
+
+extern sFONT Font24;
+extern sFONT Font20;
+extern sFONT Font16;
+extern sFONT Font12;
+extern sFONT Font8;
+extern const unsigned char Font16_Table[];
+
+#ifdef __cplusplus
+}
+#endif
+  
+#endif /* __FONTS_H */
+
