+++ src/source_image.h
@@ -1,0 +1,11 @@
+#ifndef SOURCE_IMAGE_H_
+#define SOURCE_IMAGE_H_
+#include <avr/pgmspace.h>
+/*******************
+ADD IMAGES HERE TO PROGMEM
+IMAGES ARE ACCEPTED AS HEX.
+************************/
+
+//extern PROGMEM const unsigned char gImage_70X70[];
+
+#endif
